+++ lib/commands.ps1
@@ -35,14 +35,3 @@ function Invoke-ScoopCommand {
 
     & (command_path $cmd) @arguments
 }
-
-#region Deprecated
-function exec($cmd, $arguments) {
-    # Show-DeprecatedWarning $MyInvocation 'Invoke-ScoopCommand'
-    # Invoke-ScoopCommand $cmd $arguments
-
-    $cmd_path = command_path $cmd
-
-    & $cmd_path @arguments
-}
-#endregion Deprecated