+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 0.6-pre2
 
+- Remove deprecated functions from code-base
 - **binaries**: Indicate binary execution errors with exit codes
 - Git operations with custom wrapper are now executable under Unix-like systems
 - **auto-pr**